set up navigation and viewmodels for recipe lists

+++ app/src/main/java/com/carvajal/lucas/supertaster/data/CookbookRecipe.kt
@@ -15,5 +15,5 @@ import androidx.room.PrimaryKey
 data class CookbookRecipe(
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "cookbook_id") val cookbookId: Long,
-    @ColumnInfo(name = "recipe_id") val recipeId: Int,
+    @ColumnInfo(name = "recipe_id") val recipeId: Long,
 )
